Corrected some `out` parameter types in the `TryGetLogical` methods of floating point reps
All the types matched the types in the `DoubleRep` struct, where they should be shortened to match their type.

+++ FloatingPointRep/FloatingPointRep/FloatRep.cs
@@ -248,7 +248,7 @@ public readonly record struct FloatRep
     /// <param name="IsNegative"></param>
     /// <param name="LiteralExponent"></param>
     /// <param name="LiteralMantissa"></param>
-    public void Deconstruct(out bool IsNegative, out ushort LiteralExponent, out ulong LiteralMantissa)
+    public void Deconstruct(out bool IsNegative, out byte LiteralExponent, out uint LiteralMantissa)
     {
         IsNegative = this.IsNegative;
         LiteralExponent = this.LiteralExponent;
